requirements.txt
bleach==3.2.1
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.2.3
h5py==2.10.0
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.5
jsonschema==2.6.0
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.2.6
jupyterlab-server==1.0.0
Keras==1.2.2
kiwisolver==1.1.0
MarkupSafe==1.0
matplotlib==3.0.3
mistune==0.8.3
nbconvert==5.5.0
nbformat==5.0.8
nltk==3.5
notebook==5.6.0
numpy==1.18.5
packaging==20.8
pandas==0.23.4
pandocfilters==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.9.0
prompt-toolkit==1.0.15
protobuf==3.14.0
ptyprocess==0.7.0
pydot==1.4.1
pydot-ng==2.0.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
pyzmq==17.1.2
regex==2018.8.29
scipy==1.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
tensorflow==0.12.1
terminado==0.8.1
testpath==0.4.4
Theano==1.0.5
tornado==5.1.1
tqdm==4.55.1
traitlets==4.3.2
wcwidth==0.2.5
webencodings==0.5.1
